Feat: 화면 전환 하면 날짜 유지 못하는 부분 개선
