+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/main/MainActivity.kt
@@ -92,8 +92,4 @@ class MainActivity : AppCompatActivity() {
         val REQUEST_CODE = 100
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
-        Log.d("PLANJDEBUG", "MainActivity onDestroy call")
-    }
 }
